Move TidyToolWindowController in Controllers folder
